+++ packages/core/src/types/index.ts
@@ -1,4 +1,5 @@
 export * from './article'
 export * from './about'
 export * from './category'
+export * from './friends'
 export * from './site-config'