+++ themes/vue/src/App.vue
@@ -19,6 +19,8 @@ function applyRouteTitle (): void {
     document.title = `归档 - ${site}`
   } else if (name === 'about') {
     document.title = `关于我 - ${site}`
+  } else if (name === 'friends') {
+    document.title = `友链 - ${site}`
   } else if (name === 'search') {
     document.title = `搜索 - ${site}`
   } else {